fix: clang-tidy stuff

+++ src/Core/Type/Datatype/DtDouble.cpp
@@ -7,7 +7,7 @@ namespace Apalinea::Core::Type::Datatype {
         if (other.getIdentifier() == this->getIdentifier()) {//Double
             return new DtDouble(tmpData + (static_cast<const DtDouble &>(other)).toDouble());
         } else if (other.getIdentifier() == DtUInt64::IDENTIFIER) {//U64
-            return new DtDouble(tmpData + (static_cast<const DtUInt64 &>(other)).toUInt64());
+            return new DtDouble(tmpData + static_cast<double>((static_cast<const DtUInt64 &>(other)).toUInt64()));
         } else if (other.getIdentifier() == DtUInt32::IDENTIFIER) {//U32
             return new DtDouble(tmpData + (static_cast<const DtUInt32 &>(other)).toUInt32());
         } else if (other.getIdentifier() == DtUInt16::IDENTIFIER) {//U16
@@ -15,7 +15,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtUInt8::IDENTIFIER) {//U8
             return new DtDouble(tmpData + (static_cast<const DtUInt8 &>(other)).toUInt8());
         } else if (other.getIdentifier() == DtInt64::IDENTIFIER) {//S64
-            return new DtDouble(tmpData + (static_cast<const DtInt64 &>(other)).toInt64());
+            return new DtDouble(tmpData + static_cast<double>((static_cast<const DtInt64 &>(other)).toInt64()));
         } else if (other.getIdentifier() == DtInt32::IDENTIFIER) {//S32
             return new DtDouble(tmpData + (static_cast<const DtInt32 &>(other)).toInt32());
         } else if (other.getIdentifier() == DtInt16::IDENTIFIER) {//S16
@@ -23,7 +23,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtInt8::IDENTIFIER) {//S8
             return new DtDouble(tmpData + (static_cast<const DtInt8 &>(other)).toInt8());
         } else if (other.getIdentifier() == DtSizeT::IDENTIFIER) {//SizeT
-            return new DtDouble(tmpData + (static_cast<const DtSizeT &>(other)).toSizeT());
+            return new DtDouble(tmpData + static_cast<double>((static_cast<const DtSizeT &>(other)).toSizeT()));
         } else if (other.getIdentifier() == DtBool::IDENTIFIER) {//Bool
             return new DtDouble(tmpData + (static_cast<const DtBool &>(other)).toBool());
         } else if(other.getIdentifier() == DtInt::IDENTIFIER) {//Int
@@ -41,7 +41,7 @@ namespace Apalinea::Core::Type::Datatype {
         if (other.getIdentifier() == this->getIdentifier()) {//Double
             return new DtDouble(tmpData - (static_cast<const DtDouble &>(other)).toDouble());
         } else if (other.getIdentifier() == DtUInt64::IDENTIFIER) {//U64
-            return new DtDouble(tmpData - (static_cast<const DtUInt64 &>(other)).toUInt64());
+            return new DtDouble(tmpData - static_cast<double>((static_cast<const DtUInt64 &>(other)).toUInt64()));
         } else if (other.getIdentifier() == DtUInt32::IDENTIFIER) {//U32
             return new DtDouble(tmpData - (static_cast<const DtUInt32 &>(other)).toUInt32());
         } else if (other.getIdentifier() == DtUInt16::IDENTIFIER) {//U16
@@ -49,7 +49,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtUInt8::IDENTIFIER) {//U8
             return new DtDouble(tmpData - (static_cast<const DtUInt8 &>(other)).toUInt8());
         } else if (other.getIdentifier() == DtInt64::IDENTIFIER) {//S64
-            return new DtDouble(tmpData - (static_cast<const DtInt64 &>(other)).toInt64());
+            return new DtDouble(tmpData - static_cast<double>((static_cast<const DtInt64 &>(other)).toInt64()));
         } else if (other.getIdentifier() == DtInt32::IDENTIFIER) {//S32
             return new DtDouble(tmpData - (static_cast<const DtInt32 &>(other)).toInt32());
         } else if (other.getIdentifier() == DtInt16::IDENTIFIER) {//S16
@@ -57,7 +57,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtInt8::IDENTIFIER) {//S8
             return new DtDouble(tmpData - (static_cast<const DtInt8 &>(other)).toInt8());
         } else if (other.getIdentifier() == DtSizeT::IDENTIFIER) {//SizeT
-            return new DtDouble(tmpData - (static_cast<const DtSizeT &>(other)).toSizeT());
+            return new DtDouble(tmpData - static_cast<double>((static_cast<const DtSizeT &>(other)).toSizeT()));
         } else if (other.getIdentifier() == DtBool::IDENTIFIER) {//Bool
             return new DtDouble(tmpData - (static_cast<const DtBool &>(other)).toBool());
         } else if(other.getIdentifier() == DtInt::IDENTIFIER) {//Int
