+++ src/Core/Type/Datatype/DtInt64.cpp
@@ -7,7 +7,7 @@ namespace Apalinea::Core::Type::Datatype {
         if (other.getIdentifier() == this->getIdentifier()) {//S64
             return new DtInt64(tmpData + (static_cast<const DtInt64 &>(other)).toInt64());
         } else if (other.getIdentifier() == DtUInt64::IDENTIFIER) {//U64
-            return new DtInt64(tmpData + (static_cast<const DtUInt64 &>(other)).toUInt64());
+            return new DtInt64(static_cast<int64_t>(tmpData + (static_cast<const DtUInt64 &>(other)).toUInt64()));
         } else if (other.getIdentifier() == DtUInt32::IDENTIFIER) {//U32
             return new DtInt64(tmpData + (static_cast<const DtUInt32 &>(other)).toUInt32());
         } else if (other.getIdentifier() == DtUInt16::IDENTIFIER) {//U16
@@ -23,7 +23,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtInt::IDENTIFIER) {//Int
             return new DtInt64(tmpData + (static_cast<const DtInt &>(other)).toInt());
         } else if (other.getIdentifier() == DtSizeT::IDENTIFIER) {//SizeT
-            return new DtInt64(tmpData + (static_cast<const DtSizeT &>(other)).toSizeT());
+            return new DtInt64(static_cast<int64_t>(tmpData + (static_cast<const DtSizeT &>(other)).toSizeT()));
         } else if (other.getIdentifier() == DtBool::IDENTIFIER) {//Bool
             return new DtInt64(tmpData + (static_cast<const DtBool &>(other)).toBool());
         } else if(other.getIdentifier() == DtFloat::IDENTIFIER) {//Float
@@ -41,7 +41,7 @@ namespace Apalinea::Core::Type::Datatype {
         if (other.getIdentifier() == this->getIdentifier()) {//S64
             return new DtInt64(tmpData - (static_cast<const DtInt64 &>(other)).toInt64());
         } else if (other.getIdentifier() == DtUInt64::IDENTIFIER) {//U64
-            return new DtInt64(tmpData - (static_cast<const DtUInt64 &>(other)).toUInt64());
+            return new DtInt64(static_cast<int64_t>(tmpData - (static_cast<const DtUInt64 &>(other)).toUInt64()));
         } else if (other.getIdentifier() == DtUInt32::IDENTIFIER) {//U32
             return new DtInt64(tmpData - (static_cast<const DtUInt32 &>(other)).toUInt32());
         } else if (other.getIdentifier() == DtUInt16::IDENTIFIER) {//U16
@@ -57,7 +57,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtInt::IDENTIFIER) {//Int
             return new DtInt64(tmpData - (static_cast<const DtInt &>(other)).toInt());
         } else if (other.getIdentifier() == DtSizeT::IDENTIFIER) {//SizeT
-            return new DtInt64(tmpData - (static_cast<const DtSizeT &>(other)).toSizeT());
+            return new DtInt64(static_cast<int64_t>(tmpData - (static_cast<const DtSizeT &>(other)).toSizeT()));
         } else if (other.getIdentifier() == DtBool::IDENTIFIER) {//Bool
             return new DtInt64(tmpData - (static_cast<const DtBool &>(other)).toBool());
         } else if(other.getIdentifier() == DtFloat::IDENTIFIER) {//Float
@@ -75,7 +75,7 @@ namespace Apalinea::Core::Type::Datatype {
         if (other.getIdentifier() == this->getIdentifier()) {//S64
             return new DtInt64(tmpData * (static_cast<const DtInt64 &>(other)).toInt64());
         } else if (other.getIdentifier() == DtUInt64::IDENTIFIER) {//U64
-            return new DtInt64(tmpData * (static_cast<const DtUInt64 &>(other)).toUInt64());
+            return new DtInt64(static_cast<int64_t>(tmpData * (static_cast<const DtUInt64 &>(other)).toUInt64()));
         } else if (other.getIdentifier() == DtUInt32::IDENTIFIER) {//U32
             return new DtInt64(tmpData * (static_cast<const DtUInt32 &>(other)).toUInt32());
         } else if (other.getIdentifier() == DtUInt16::IDENTIFIER) {//U16
@@ -91,7 +91,7 @@ namespace Apalinea::Core::Type::Datatype {
         } else if (other.getIdentifier() == DtInt::IDENTIFIER) {//Int
             return new DtInt64(tmpData * (static_cast<const DtInt &>(other)).toInt());
         } else if (other.getIdentifier() == DtSizeT::IDENTIFIER) {//SizeT
-            return new DtInt64(tmpData * (static_cast<const DtSizeT &>(other)).toSizeT());
+           return new DtInt64(static_cast<int64_t>(tmpData * (static_cast<const DtSizeT &>(other)).toSizeT()));
         } else if (other.getIdentifier() == DtBool::IDENTIFIER) {//Bool
             return new DtInt64(tmpData * (static_cast<const DtBool &>(other)).toBool());
         } else if(other.getIdentifier() == DtFloat::IDENTIFIER) {//Float
@@ -141,27 +141,27 @@ namespace Apalinea::Core::Type::Datatype {
         int64_t tmpData = this->toInt64();
 
         if (other.getIdentifier() == this->getIdentifier()) {//S64
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtInt64 &>(other)).toInt64()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtInt64 &>(other)).toInt64())));
         } else if (other.getIdentifier() == DtUInt64::IDENTIFIER) {//U64
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtUInt64 &>(other)).toUInt64()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtUInt64 &>(other)).toUInt64())));
         } else if (other.getIdentifier() == DtUInt32::IDENTIFIER) {//U32
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtUInt32 &>(other)).toUInt32()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtUInt32 &>(other)).toUInt32())));
         } else if (other.getIdentifier() == DtUInt16::IDENTIFIER) {//U16
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtUInt16 &>(other)).toUInt16()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtUInt16 &>(other)).toUInt16())));
         } else if (other.getIdentifier() == DtUInt8::IDENTIFIER) {//U8
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtUInt8 &>(other)).toUInt8()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtUInt8 &>(other)).toUInt8())));
         } else if (other.getIdentifier() == DtInt32::IDENTIFIER) {//S32
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtInt32 &>(other)).toInt32()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtInt32 &>(other)).toInt32())));
         } else if (other.getIdentifier() == DtInt16::IDENTIFIER) {//S16
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtInt16 &>(other)).toInt16()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtInt16 &>(other)).toInt16())));
         } else if (other.getIdentifier() == DtInt8::IDENTIFIER) {//S8
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtInt8 &>(other)).toInt8()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtInt8 &>(other)).toInt8())));
         } else if (other.getIdentifier() == DtInt::IDENTIFIER) {//Int
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtInt &>(other)).toInt()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtInt &>(other)).toInt())));
         } else if (other.getIdentifier() == DtSizeT::IDENTIFIER) {//SizeT
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtSizeT &>(other)).toSizeT()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtSizeT &>(other)).toSizeT())));
         } else if (other.getIdentifier() == DtBool::IDENTIFIER) {//Bool
-            return new DtInt64(std::pow(tmpData,(static_cast<const DtBool &>(other)).toBool()));
+            return new DtInt64(static_cast<int64_t>(std::pow(tmpData,(static_cast<const DtBool &>(other)).toBool())));
         } else if(other.getIdentifier() == DtFloat::IDENTIFIER) {//Float
             return new DtDouble(std::pow(tmpData,(static_cast<const DtFloat &>(other)).toFloat()));
         } else if(other.getIdentifier() == DtDouble::IDENTIFIER) {//Double
